refactor: Update naming and component structure in navigation
Rename footerNavigation to sidebarNavigationWrapper for clearer
naming conventions. Addition of sidebarTransition component leads to
the splitting of LayoutFooter component, enhancing the clarity in
the utilization of each component.
